requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docker==4.2.0
idna==2.9
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
